requirements.txt
click==8.1.7
gmsh==4.13.0
iapws==1.5.4
matplotlib==3.8.4
meshio==5.3.5
mpmath==1.3.0
numba==0.59.1
numpy==1.26.4
open-darts==1.1.4
opmcpg==0.0.3
pandas==2.2.2
pyevtk==1.6.0
pytest
pyvista==0.43.8
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.4
vtk==9.3.0
xarray==2024.5.0
